python-realtime-asr/requirement.txt
websocket_client
